Fixes for IE7
Issues stemmed from a lack of doc type - adding a doc type caused i.e.
issues to affect all browsers.  Most issues were related to a lack of px
on positional css.  Also found and fixed an issue relating to the
treatment of conditional events via backbone - this is now sorted and
probably the neatest way of doing this we can currently think of.
Tested in Firefox 5, Chrome, Opera, Safari, IE7/8/9